Refactoring to separate database stuff
